Follow upstream changes to quit64 as size
follow-up to a1cf31faf1a8aeebf2786715590cf3bb8a5cd472

+++ libdiscover/backends/AlpineApkBackend/AlpineApkResource.cpp
@@ -72,9 +72,9 @@ QString AlpineApkResource::comment()
     return m_pkg.description;
 }
 
-int AlpineApkResource::size()
+quint64 AlpineApkResource::size()
 {
-    return static_cast<int>(m_pkg.size);
+    return m_pkg.size;
 }
 
 QUrl AlpineApkResource::homepage()
